DeleteBatchHandler: check if code belongs to user in queue
